+++ Menu-o-Matic/scripts/lauchDialogFromNodes.groovy
@@ -3,4 +3,9 @@ import edofro.menuomatic.PackMenu as PM
 
 
 def (MD, resp) = PM.getMDfromNodes(node)
-if(MD) LD.showDialogFromMD(MD)
+if(MD) {
+    LD.showDialogFromMD(MD)
+} else {
+    def msg = 'MDI: No command or groovy node found as descendant of the selected node'
+    c.statusInfo = msg
+}